Fixed communication with daemon, improved debug
* execTab.cpp, radeon_profile.cpp, uiEvents.h, uiEvents.cpp: When
starting an interface, Qt calls connectSlotsByName which parses the
functions and finds any slot named with the pattern on_objectName_signal,
and tries to attach automatically to the signal 'signal' from the object
'objectName'. In our case it found the slot on_btn_backToProfiles_clicked
but didn't find 'btn_backToProfiles' since we create it at runtime, so
each time the program started an it logged this error:
"QMetaObject::connectSlotsByName: No matching signal for
on_btn_backToProfiles_clicked()"; by renaming the slot as
'btnBackToProfilesClicked()', the error is gone.
* dxorg.h, dxorg.cpp(34, 140): Added SHARED_MEM_SIZE constant.
* dxorg.cpp(34-37): Fixed sharedMem creation checks and debug.
* dxorg.cpp(64, 524, 557, 571, 588): Now each part of the communication
is separated, need the last commit of radeon-profile-daemon to work well.
* dxorg.cpp (114-148): Improved the shared memory managment, now before
reading the memory pointer it is checked to prevent segfaults, and also
QByteArray now manages all the data movements through
QByteArray::fromRawData and QFile::readAll.