amdgpuIoctlHandler: fix GPU usage According to amdgpu source code, AMDGPU_INFO_SENSOR_GPU_LOAD is an integer: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/drivers/gpu/drm/amd/amdgpu/amdgpu_kms.c#n854
In getSensorValue(), data (a float*) is casted as uint64_t, and value is incorrect due to cast.
Otherwise, amdgpuIoctlHandler::getGpuUsage() parameter type cannot be changed because of radeonIoctlHandler::getGpuUsage(), who is really a floating value.

So, we need to pass integer to getSensorValue() and keep float in getGpuUsage(). Using a temporary variable to cast data as an integer seems to be a good compromise.
Resolves #70