handle null chars in dxorg.cpp:getValueFromSysFsFile()
QString truncates strings after the first occurance of a null character. If a sysfs file contains null characters in the middle of the contents, the output of getValueFromSysFsFile() is incomplete.

This replaces all occurances of null chars in the QByteArray returned from QFile::readAll() before converting it to a QString.
This also fixes segfaults on startup when incomplete pp_od_clk_voltage tables are being parsed.