Add bounds check to move left/right buttons in topbar cfg dialog
Make sure an item is actually selected and can be moved in the desired
direction. I.e. don't try to move left if it is already the first item
and don't move right if it is the last.

Without this bounds check clicking on move left/right leads to crashes
due to out of range acccesses on ui->listWidget.