Add driver level support for multiple temperature sensors
When temperature sensors via the HWMON backend are supported, detect all
available temperature sensors.

If the labels provided by hwmon are known, we assign known instance ids
to the ValueIDs of the sensor data. As of now, this includes edge,
junction and mem temperature, which will be assigned the intance of
T_EDGE, T_JUNCTION and T_MEM for all TEMPERATURE_* values.

For those instances, getNameOfValueID() now returns the appropriate
localized strings for UI labels.

For all sensors detected with unknown labels, a free instance id is
assigned, the label provided by HWMON is used to construct labels for
the UI, as returned by getNameOfValueID().