Make hysteresis a per fan profile setting
Hysteresis was a global setting, applied independently of fan profile.
This is somewhat confusing, as one could expect it to be an aspect of a
fan profile, since one sets it in the fan profile tab.

This adds a new FanProfile struct that can store additional settings
besides the temperature steps and adds hysteresis as one of those
settings.

The global hysteresis value stored in older radeon-profile-settings will
still be read and applied to all loaded profiles. Once new config files
are written, this values is moved to the xml file and stored as attributes
of each fan profile respectively.