allow selecting temperature sensor in fan profiles
Adds a new combo box to the fan profile tab to allow the user to select
a temperature sensor as data source for this fan curve.

radeon_profile::adjustFanSpeed() will now read the current temperature
of the temperature sensor selected in the currently active fan profile.

This setting is a per fan profile setting. It is stored in the
newly added FanProfile struct and written to and read from the xml config file
for each fan profile. By default, the edge temperature sensor is
selected. This also applies when an older config file is read without
the new config attribute.